+++ config/vuedata.php
@@ -2,5 +2,5 @@
 
 // config for Itstudioat/Vuedata
 return [
-    'version' => '1.0.3'
+    'version' => '1.0.4'
 ];