+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to `vuedata` will be documented in this file.
 
+## V1.0.4
+[
+    'success' => false,
+    'error' => 'File not found'
+]:
+[
+    'success' => true,
+    'data' => [...],
+]
+
+[
+    'success' => false,
+    'error' => 'File not found'
+]
+
+
 ## V1.0.3
 Dont write the Debug-Logfile